+++ keepalived/include/vrrp_scheduler.h
@@ -65,7 +65,7 @@ do {						\
 extern void vrrp_dispatcher_release(vrrp_data_t *);
 extern int vrrp_dispatcher_init(thread_t *);
 extern int vrrp_lower_prio_gratuitous_arp_thread(thread_t *);
-extern void vrrp_set_effective_priority(vrrp_t *, int);
+extern void vrrp_set_effective_priority(vrrp_t *, uint8_t);
 extern int vrrp_arp_thread(thread_t *);
 
 #endif
